Using supers now
Yeah, so I didn't like how it was set up before so I added the base constructor in the abstract event for easy extension for other classes. I gotta add the company in to here because it's not a part of the 'default' event which is personalEvent

P.S. Also, this is my favorite class. It really is something I never could have imagined myself doing at the start of OOP, and the same goes for every class, but this one just feels so simple and so good. This file, in context to everything else in this project, feels like a great show of how much I've learned